# Use concurrent thread pool to download sidecar items